% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancillary.R
\name{aaf}
\alias{aaf}
\title{Compute the Frequency of Each Amino Acid in Each Species}
\usage{
aaf(data)
}
\arguments{
\item{data}{input data must be a dataframe (see details).}
}
\value{
A dataframe providing amino acid frequencies en the set of species. Rows correspond amino acids and columns to species.
}
\description{
Computes the frequency of each amino acid in each species.
}
\details{
Input data must be a dataframe where each row corresponds to an individual protein, and each column identifies a species. Therefore, the columns' names of this dataframe must be coherent with the names of the OTUs being analyzed.
}
\examples{
aaf(bovids)
}
\seealso{
env.sp(), otu.vector(), otu.space()
}
