% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{df2fasta}
\alias{df2fasta}
\title{Convert Dataframe into Fasta File}
\usage{
df2fasta(df, out.file)
}
\arguments{
\item{df}{a named (both rows and cols) dataframe (see details).}

\item{out.file}{path and name of output file.}
}
\value{
A fasta file that is saved in the specified path.
}
\description{
Converts a dataframe into a fasta file.
}
\details{
The format of the df should be as follows. Each row represents a protein sequence and each column a species.
}
\examples{
\dontrun{df2fasta(df = bovids, out.file = "./example.fasta")}
}
\seealso{
fastaconc()
}
