% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envfas.R
\name{vect2tree}
\alias{vect2tree}
\title{Convert a Set of Vectors into a Tree}
\usage{
vect2tree(path, metric = "cosine", clustering = "nj")
}
\arguments{
\item{path}{path to the working directory. This directory must contain a txt file per vector and an additional txt file named vlist.txt that provides the names (one per line) of the  vector txt files.}

\item{metric}{character string indicating the metric (see metrics() to see the methods allowed).}

\item{clustering}{string indicating the clustering method, either "nj" or "upgma".}
}
\value{
a list with two elements: a distance matrix and a tree.
}
\description{
Converts a set of vectors into a tree.
}
\details{
This function computes the distance matrix and builds the corresponding tree.
}
\examples{
\dontrun{vec2tree("./data_t")}
}
\seealso{
envnj(), fastaconc(), envfascpp()
}
